+++ src/main/java/insane96mcp/enhancedai/modules/mobs/miner/MinerMobs.java
@@ -2,75 +2,83 @@ package insane96mcp.enhancedai.modules.mobs.miner;
 
 import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.data.EAIData;
+import insane96mcp.enhancedai.data.EAIDataEnum;
 import insane96mcp.enhancedai.data.EAIDataList;
 import insane96mcp.enhancedai.modules.Modules;
-import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.enhancedai.utils.GoalHelper;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
-import insane96mcp.insanelib.base.config.Blacklist;
 import insane96mcp.insanelib.base.config.Config;
-import insane96mcp.insanelib.data.IdTagMatcher;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraftforge.common.ForgeMod;
+import net.minecraftforge.event.entity.EntityAttributeModificationEvent;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.List;
 
-@LoadFeature(module = Modules.Ids.MOBS, description = "Mobs can mine blocks to reach the target. Uses offhand item to mine. Only mobs in the entity type tag enhancedai:miner_mobs/can_mine can spawn with the ability to mine and blocks in the tag enhancedai:miner_mobs/blacklist cannot be mined.")
+@LoadFeature(module = Modules.Ids.MOBS, description = "Mobs can mine blocks to reach the target. Uses offhand item to mine. Only mobs in the entity type tag enhancedai:miner_mobs/can_mine can spawn with the ability to mine and blocks in the tag enhancedai:miner_mobs/blacklist cannot be mined. This feature also adds the block reach attribute to all entities.")
 public class MinerMobs extends Feature {
 	public static final TagKey<EntityType<?>> CAN_BE_MINER = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("miner_mobs/can_mine"));
 	public static final TagKey<Block> BLOCK_BLACKLIST = TagKey.create(Registries.BLOCK, EnhancedAI.location("miner_mobs/blacklist"));
-	public static final String TIME_TO_BREAK_MULTIPLIER = EnhancedAI.RESOURCE_PREFIX + "time_to_break_multiplier";
-	public static final String TOOL_ONLY = EnhancedAI.RESOURCE_PREFIX + "tool_only";
-	public static final String PROPER_TOOL_ONLY = EnhancedAI.RESOURCE_PREFIX + "proper_tool_only";
-	public static final String ALWAYS_REQUIRE_PROPER_TOOL = EnhancedAI.RESOURCE_PREFIX + "always_require_proper_tool";
 
 	@Config(min = 0d, max = 1d, description = "Chance for a mob in the entity type tag enhancedai:can_be_miner to spawn with the miner ability")
 	public static Double minerChance = 0.07d;
-	@Config(description = "Mobs with the miner AI will mine only if they have any tool in the off-hand")
-	public static Boolean canMineWithToolOnly = true;
-	@Config(description = "Mobs with the miner AI will mine only if their off-hand tool can mine targeted blocks (e.g. zombies with shovels will not mine stone). Blocks that require no tool (e.g. planks) will be minable regardless of proper tool or not.")
-	public static Boolean canMineWithProperToolOnly = false;
-	@Config(description = "If 'Can mine with proper tool only' is enabled, mobs with the miner AI will mine blocks that don't require a tool only with the proper tool.")
-	public static Boolean alwaysRequireProperTool = false;
-	@Config(description = "Mobs with Miner AI will spawn with a Stone Pickaxe that never drops.")
-	public static Boolean equipStonePick = true;
+	@Config(description = "NONE: No item is required. ANY_TOOL: Any item that can break blocks in the off-hand is required. CORRECT_TOOL_FOR_REQUIRED: The mob is able to mine any blocks that don't require a tool, and require a tool for blocks that require it (e.g. can always mine dirt but can't mine stone). CORRECT_TOOL_FOR_ANY_BLOCK: The mob can only mine blocks if the tool is the right one for the block (e.g. can mine dirt only with a shovel).")
+	public static ToolRequirement toolRequirement = ToolRequirement.CORRECT_TOOL_FOR_REQUIRED;
 	@Config(min = -512, max = 1024, description = "Mobs can mine from the bottom of the world to this Y level.")
 	public static Integer maxY = 320;
 	@Config(min = 0, max = 128, description = "The maximum distance from the target at which the Mobs can mine. Set to 0 to always mine.")
-	public static Integer maxDistance = 0;
-	@Config(description = "Mobs with Miner AI will not be able to break tile entities")
-	public static Boolean blacklistTileEntities = false;
+	public static Integer maxTargetDistance = 0;
 	@Config(min = 0d, max = 128d, description = "Multiplier for the time a mob takes to break blocks. E.g. with this set to 2, mobs will take twice the time to mine a block.")
 	public static Double timeToBreakMultiplier = 1.25d;
-	@Config(description = "If true, the block tag `enhancedai:miner_block_blacklist` will be treated as a whitelist instead of blacklist")
+	@Config(description = "Dimensions where mobs can mine.")
+	public static List<String> dimensionWhitelist = List.of("minecraft:overworld", "minecraft:the_nether", "minecraft:the_end");
+	@Config(description = "If true, the block tag `enhancedai:miner_mobs/blacklist` will be treated as a whitelist instead of blacklist")
 	public static Boolean blockBlacklistAsWhitelist = false;
-	@Config(description = "Dimensions where mobs can't spawn with the ability to mine.")
-	public static Blacklist dimensionBlacklist = new Blacklist();
+	@Config(description = "Mobs with Miner AI will not be able to break tile entities")
+	public static Boolean blacklistTileEntities = true;
+	@Config(description = "Mobs with Miner AI will spawn with a Stone Pickaxe that never drops.")
+	public static Boolean equipStonePick = true;
 
 	public static EAIData<Boolean> MINER;
+	public static EAIDataEnum<ToolRequirement> TOOL_REQUIREMENT;
+	public static EAIData<Integer> MAX_Y;
+	public static EAIData<Integer> MAX_TARGET_DISTANCE;
+	public static EAIData<Double> TIME_TO_BREAK_MULTIPLIER;
 	public static EAIDataList<String> DIMENSION_WHITELIST;
 
 	public void init(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super.init(module, enabledByDefault, canBeDisabled);
 		MINER = EAIData.ofBool(this.createDataKey("miner"), (mob, miner) -> {
-			GoalHelper.removeGoal(mob.goalSelector, BlockBreakerGoal.class);
+			GoalHelper.removeGoal(mob.goalSelector, MineTowardsTargetGoal.class);
 			if (miner)
-				mob.goalSelector.addGoal(1, new BlockBreakerGoal(mob, maxDistance, timeToBreakMultiplier, canMineWithToolOnly, canMineWithProperToolOnly, alwaysRequireProperTool));
+				mob.goalSelector.addGoal(1, new MineTowardsTargetGoal(mob));
 		});
+		TOOL_REQUIREMENT = EAIDataEnum.of(this.createDataKey("tool_requirement"), ToolRequirement.class);
+		MAX_Y = EAIData.ofInt(this.createDataKey("max_y"));
+		MAX_TARGET_DISTANCE = EAIData.ofInt(this.createDataKey("max_target_distance"));
+		TIME_TO_BREAK_MULTIPLIER = EAIData.ofDouble(this.createDataKey("time_to_break_multiplier"));
 		DIMENSION_WHITELIST = EAIDataList.of(this.createDataKey("dimension_whitelist"), String.class);
+	}
+
+	public static void addAttribute(EntityAttributeModificationEvent event) {
+		for (EntityType<? extends LivingEntity> entityType : event.getTypes()) {
+			if (event.has(entityType, ForgeMod.BLOCK_REACH.get()))
+				continue;
+
+			event.add(entityType, ForgeMod.BLOCK_REACH.get());
+		}
 	}
 
 	//Lowest priority so other mods can set persistent data
@@ -79,40 +87,36 @@ public class MinerMobs extends Feature {
 		if (!this.isEnabled()
 		 		|| event.getLevel().isClientSide
 				|| !(event.getEntity() instanceof Mob mob)
-		 		|| !mob.getType().is(CAN_BE_MINER)
-				|| MinerMobs.isDimensionBlackOrNotWhitelisted(event.getLevel()))
+		 		|| !mob.getType().is(CAN_BE_MINER))
 			return;
 
-		CompoundTag persistentData = mob.getPersistentData();
-
-		double timeToBreakMultiplier1 = NBTUtils.getDoubleOrPutDefaultLegacy(persistentData, TIME_TO_BREAK_MULTIPLIER, timeToBreakMultiplier);
-		boolean toolOnly = NBTUtils.getBooleanOrPutDefaultLegacy(persistentData, TOOL_ONLY, canMineWithToolOnly);
-		boolean properToolOnly = NBTUtils.getBooleanOrPutDefaultLegacy(persistentData, PROPER_TOOL_ONLY, canMineWithProperToolOnly);
-		boolean properToolRequired = NBTUtils.getBooleanOrPutDefaultLegacy(persistentData, ALWAYS_REQUIRE_PROPER_TOOL, alwaysRequireProperTool);
-
-		if (equipStonePick && mob.getOffhandItem().isEmpty())
+		boolean isMiner = mob.getRandom().nextDouble() < minerChance;
+		MINER.applyIfAbsent(mob, isMiner);
+		if (isMiner && equipStonePick && mob.getOffhandItem().isEmpty())
 		{
 			mob.setItemSlot(EquipmentSlot.OFFHAND, new ItemStack(Items.STONE_PICKAXE));
 			mob.setDropChance(EquipmentSlot.OFFHAND, -1f);
 		}
-		MINER.applyIfAbsent(mob, mob.getRandom().nextDouble() < minerChance);
-		DIMENSION_WHITELIST.applyIfAbsent(mob, List.of("minecraft:overworld", "minecraft:the_nether", "minecraft:the_end"));
+		TOOL_REQUIREMENT.applyIfAbsent(mob, toolRequirement);
+		MAX_Y.applyIfAbsent(mob, maxY);
+		MAX_TARGET_DISTANCE.applyIfAbsent(mob, maxTargetDistance);
+		TIME_TO_BREAK_MULTIPLIER.applyIfAbsent(mob, timeToBreakMultiplier);
+		DIMENSION_WHITELIST.applyIfAbsent(mob, dimensionWhitelist);
 	}
 
-	public static boolean isDimensionBlackOrNotWhitelisted(Level level) {
-		//Check for black/whitelist
-		boolean isInWhitelist = false;
-		boolean isInBlacklist = false;
-		for (IdTagMatcher blacklistEntry : dimensionBlacklist.blacklist) {
-			if (blacklistEntry.location.equals(level.dimension().location())) {
-				if (!dimensionBlacklist.blacklistAsWhitelist)
-					isInBlacklist = true;
-				else
-					isInWhitelist = true;
-				break;
-			}
+	public static boolean isValidDimension(Mob mob) {
+		List<String> dimensionWhitelist = DIMENSION_WHITELIST.get(mob);
+		for (String dimension : dimensionWhitelist) {
+			if (dimension.equals(mob.level().dimension().location().toString()))
+				return true;
 		}
+		return false;
+	}
 
-		return isInBlacklist || (!isInWhitelist && dimensionBlacklist.blacklistAsWhitelist);
+	public enum ToolRequirement {
+		NONE,
+		ANY_TOOL,
+		CORRECT_TOOL_FOR_REQUIRED,
+		CORRECT_TOOL_FOR_ANY_BLOCK
 	}
 }